Updated version after release

+++ src/main/java/services/DataService.java
@@ -154,7 +154,7 @@ public class DataService {
                 eq(Const.UID, itemUid)),
                     Updates.set(Const.CATEGORY_UID, trash.getUid()));
 
-        if (updateResult.getModifiedCount() == 1) {
+        if (updateResult.getModifiedCount() == 1 && StringUtils.isNotBlank(item.getMediaUid())) {
             mediaService.delete(item.getMediaUid(), userUid);
         }
 
@@ -357,7 +357,9 @@ public class DataService {
         var user = findUserByUid(userUid);
         if (user != null && user.getPassword().equals(CodecUtils.hashArgon2(password, user.getSalt()))) {
             List<Item> items = datastore.findAll(Item.class, eq(Const.USER_UID, userUid), Sorts.ascending(Const.USER_UID));
-            items.forEach(item -> mediaService.delete(item.getMediaUid(), userUid));
+            items.stream()
+                    .filter(item -> StringUtils.isNotBlank(item.getMediaUid()))
+                    .forEach(item -> mediaService.delete(item.getMediaUid(), userUid));
 
             DeleteResult deleteCategories = datastore.query(Category.class).deleteMany(eq(Const.USER_UID, userUid));
             DeleteResult deleteItems = datastore.query(Item.class).deleteMany(eq(Const.USER_UID, userUid));
